requirements.txt
tqdm
ninja
accelerate
transformers
rdkit
deepspeed
ray[tune]
wandb
rdkit
snntorch 
optuna 
optuna-dashboard
zeus-ml
argparse==1.4.0 ; python_version >= "3.8" and python_version < "3.11"
asteval==0.9.31 ; python_version >= "3.8" and python_version < "3.11"
certifi==2023.11.17 ; python_version >= "3.8" and python_version < "3.11"
charset-normalizer==3.3.2 ; python_version >= "3.8" and python_version < "3.11"
cmake==3.27.7 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "3.11" and sys_platform == "win32"
contourpy==1.1.1 ; python_version >= "3.8" and python_version < "3.11"
cycler==0.12.1 ; python_version >= "3.8" and python_version < "3.11"
exceptiongroup==1.2.0 ; python_version >= "3.8" and python_version < "3.11"
filelock==3.13.1 ; python_version >= "3.8" and python_version < "3.11"
fonttools==4.45.1 ; python_version >= "3.8" and python_version < "3.11"
h5py==3.10.0 ; python_version >= "3.8" and python_version < "3.11"
idna==3.7 ; python_version >= "3.8" and python_version < "3.11"
importlib-resources==6.1.1 ; python_version >= "3.8" and python_version < "3.10"
iniconfig==2.0.0 ; python_version >= "3.8" and python_version < "3.11"
jinja2==3.1.3 ; python_version >= "3.8" and python_version < "3.11"
joblib==1.3.2 ; python_version >= "3.8" and python_version < "3.11"
kiwisolver==1.4.5 ; python_version >= "3.8" and python_version < "3.11"
lava-nc @ git+https://github.com/lava-nc/lava.git@06110649c1d528bf61b6dd2266e34100a813e3fd ; python_version >= "3.8" and python_version < "3.11"
linecache2==1.0.0 ; python_version >= "3.8" and python_version < "3.11"
lit==17.0.5 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
markupsafe==2.1.3 ; python_version >= "3.8" and python_version < "3.11"
matplotlib==3.7.4 ; python_version >= "3.8" and python_version < "3.11"
mpmath==1.3.0 ; python_version >= "3.8" and python_version < "3.11"
networkx==2.8.7 ; python_version >= "3.8" and python_version < "3.11"
ninja==1.11.1.1 ; python_version >= "3.8" and python_version < "3.11"
numpy==1.24.4 ; python_version >= "3.8" and python_version < "3.11"
nvidia-cublas-cu11==11.10.3.66 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
nvidia-cuda-cupti-cu11==11.7.101 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
nvidia-cuda-nvrtc-cu11==11.7.99 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
nvidia-cuda-runtime-cu11==11.7.99 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
nvidia-cudnn-cu11==8.5.0.96 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
nvidia-cufft-cu11==10.9.0.58 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
nvidia-curand-cu11==10.2.10.91 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
nvidia-cusolver-cu11==11.4.0.1 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
nvidia-cusparse-cu11==11.7.4.91 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
nvidia-nccl-cu11==2.14.3 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
nvidia-nvtx-cu11==11.7.91 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
opencv-python-headless==4.5.5.64 ; python_version >= "3.8" and python_version < "3.11"
packaging==23.2 ; python_version >= "3.8" and python_version < "3.11"
pillow==10.3.0 ; python_version >= "3.8" and python_version < "3.11"
pluggy==1.3.0 ; python_version >= "3.8" and python_version < "3.11"
pyparsing==3.1.1 ; python_version >= "3.8" and python_version < "3.11"
pytest==7.4.3 ; python_version >= "3.8" and python_version < "3.11"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "3.11"
requests==2.31.0 ; python_version >= "3.8" and python_version < "3.11"
scikit-learn==1.3.2 ; python_version >= "3.8" and python_version < "3.11"
scipy==1.10.1 ; python_version >= "3.8" and python_version < "3.11"
setuptools==69.0.2 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.8" and python_version < "3.11"
sympy==1.12 ; python_version >= "3.8" and python_version < "3.11"
threadpoolctl==3.2.0 ; python_version >= "3.8" and python_version < "3.11"
tomli==2.0.1 ; python_version >= "3.8" and python_version < "3.11"
torch==2.0.0 ; python_version >= "3.8" and python_version < "3.11"
torchvision==0.15.1 ; python_version >= "3.8" and python_version < "3.11"
traceback2==1.4.0 ; python_version >= "3.8" and python_version < "3.11"
triton==2.0.0 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
typing-extensions==4.8.0 ; python_version >= "3.8" and python_version < "3.11"
unittest2==1.1.0 ; python_version >= "3.8" and python_version < "3.11"
urllib3==2.1.0 ; python_version >= "3.8" and python_version < "3.11"
wheel==0.42.0 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.8" and python_version < "3.11"
zipp==3.17.0 ; python_version >= "3.8" and python_version < "3.10"